Split commands per file.
